Create UIActivity subclass and move shared code
