+++ Client/Frontend/Share/CopyLinkActivity.swift
@@ -4,40 +4,14 @@
 
 import Foundation
 
-class CopyLinkActivity: UIActivity {
-    var appActivityType: CustomActivityAction
-    var url: URL
+class CopyLinkActivity: CustomAppActivity {
 
-    override var activityTitle: String? {
-        return appActivityType.title
-    }
-
-    override var activityImage: UIImage? {
-        return appActivityType.image
-    }
-
-    override var activityType: UIActivity.ActivityType {
-        return appActivityType.actionType
-    }
-
-    override class var activityCategory: UIActivity.Category {
-        return .action
-    }
-
-    init(activityType: CustomActivityAction, url: URL) {
-        appActivityType = activityType
-        self.url = url
-        super.init()
-    }
-
-    // Send to device is only available for URL that are files
+    // Copy link is only available for URL that are not files
     override func canPerform(withActivityItems activityItems: [Any]) -> Bool {
         return !url.isFile
     }
 
-    override func prepare(withActivityItems activityItems: [Any]) {
-//        self.activityItems = activityItems
-    }
+    override func prepare(withActivityItems activityItems: [Any]) {}
 
     override func perform() {
         UIPasteboard.general.string = url.absoluteString
