+++ Client/Frontend/Share/SendToDeviceActivity.swift
@@ -5,41 +5,14 @@
 import UIKit
 import Shared
 
-class SendToDeviceActivity: UIActivity {
-    var appActivityType: CustomActivityAction
-    var activityItems = [Any]()
-    var url: URL
-
-    override var activityTitle: String? {
-        return appActivityType.title
-    }
-
-    override var activityImage: UIImage? {
-        return appActivityType.image
-    }
-
-    override var activityType: UIActivity.ActivityType {
-        return appActivityType.actionType
-    }
-
-    override class var activityCategory: UIActivity.Category {
-        return .action
-    }
-
-    init(activityType: CustomActivityAction, url: URL) {
-        appActivityType = activityType
-        self.url = url
-        super.init()
-    }
+class SendToDeviceActivity: CustomAppActivity {
 
     // Send to device is only available for URL that are files
     override func canPerform(withActivityItems activityItems: [Any]) -> Bool {
         return !url.isFile
     }
 
-    override func prepare(withActivityItems activityItems: [Any]) {
-        self.activityItems = activityItems
-    }
+    override func prepare(withActivityItems activityItems: [Any]) {}
 
     override func perform() {
         activityDidFinish(true)
